fix: address clippy warnings for --no-default-features CI
- manual_flatten in list_workspaces / list_sessions_for
- clone_on_copy for ExecApprovalMode
- collapsible_if for nested desktop checks
- unnecessary_cast usize
- allow too_many_arguments for refactored handlers

+++ src/app_state.rs
@@ -758,7 +758,7 @@ impl App {
                 self.needs_redraw = true;
             }
             KeyCode::Down | KeyCode::Char('j') => {
-                if self.selected_agent_mode_idx < (self.agent_modes.len() - 1) as usize {
+                if self.selected_agent_mode_idx < self.agent_modes.len() - 1 {
                     self.selected_agent_mode_idx += 1;
                 }
                 self.needs_redraw = true;
